gl: Refactor buffer object code
